+++ wsgidav/http_authenticator.py
@@ -136,9 +136,9 @@ class SimpleDomainController(object):
 class HTTPAuthenticator(BaseMiddleware):
     """WSGI Middleware for basic and digest authenticator."""
 
-    def __init__(self, application, config):
-        self._verbose = config.get("verbose", 2)
-        self._application = application
+    def __init__(self, wsgidav_app, next_app, config):
+        super(HTTPAuthenticator, self).__init__(wsgidav_app, next_app, config)
+        self._verbose = config.get("verbose", 3)
         self._user_mapping = config.get("user_mapping", {})
         self._domaincontroller = config.get(
             "domaincontroller") or WsgiDAVDomainController(self._user_mapping)
@@ -158,7 +158,7 @@ class HTTPAuthenticator(BaseMiddleware):
         if self._domaincontroller.__class__.__name__ == wdcName:
             if self._authacceptdigest or self._authdefaultdigest or not self._authacceptbasic:
                 _logger.warn(
-                    "WARNING: {} requires basic authentication.\n\tSet acceptbasic=True, "
+                    "{} requires basic authentication.\n\tSet acceptbasic=True, "
                     "acceptdigest=False, defaultdigest=False".format(wdcName))
 
     def getDomainController(self):
@@ -184,7 +184,7 @@ class HTTPAuthenticator(BaseMiddleware):
             _logger.debug("No authorization required for realm '{}'".format(realmname))
             environ["http_authenticator.realm"] = realmname
             environ["http_authenticator.username"] = ""
-            return self._application(environ, start_response)
+            return self.next_app(environ, start_response)
 
         if self._trusted_auth_header and environ.get(self._trusted_auth_header):
             # accept a username that was injected by a trusted upstream server
@@ -192,7 +192,7 @@ class HTTPAuthenticator(BaseMiddleware):
                     self._trusted_auth_header, environ.get(self._trusted_auth_header), realmname))
             environ["http_authenticator.realm"] = realmname
             environ["http_authenticator.username"] = environ.get(self._trusted_auth_header)
-            return self._application(environ, start_response)
+            return self.next_app(environ, start_response)
 
         if "HTTP_AUTHORIZATION" in environ:
             authheader = environ["HTTP_AUTHORIZATION"]
@@ -257,7 +257,7 @@ class HTTPAuthenticator(BaseMiddleware):
         if self._domaincontroller.authDomainUser(realmname, username, password, environ):
             environ["http_authenticator.realm"] = realmname
             environ["http_authenticator.username"] = username
-            return self._application(environ, start_response)
+            return self.next_app(environ, start_response)
         return self.sendBasicAuthResponse(environ, start_response)
 
     def sendDigestAuthResponse(self, environ, start_response):
@@ -422,7 +422,7 @@ class HTTPAuthenticator(BaseMiddleware):
 
         environ["http_authenticator.realm"] = realmname
         environ["http_authenticator.username"] = req_username
-        return self._application(environ, start_response)
+        return self.next_app(environ, start_response)
 
     def computeDigestResponse(
             self, username, realm, password, method, uri, nonce, cnonce, qop, nc):